requirements/requirements.txt
pandas>=1.5.3
numpy>=1.24.3
matplotlib>=3.7.1
seaborn>=0.12.2
jupyter>=1.0.0
openpyxl>=3.1.2
sqlalchemy>=2.0.20
mysql-connector-python>=8.1.0
python-dotenv>=1.0.0
plotly>=5.16.1
dash>=2.13.0
scikit-learn>=1.3.0
pymysql>=1.1.0 